+++ backend/src/main/java/com/jaba/vgl/repositories/impl/ReviewRepositoryImpl.java
@@ -26,7 +26,7 @@ public interface ReviewRepositoryImpl extends ReviewRepository, ReviewRepository
             "r.rating = :rating " +
             "WHERE r.id = :id")
     @Override
-    void updateReview(Long id, Long gameId, String title, String text, Float rating);
+    void updateReview(Long id, Long gameId, String title, String text, Integer rating);
 
     @Override
     @Query("SELECT r.id FROM Review r WHERE r.title = ?1")
